// backend/src/types/professional/professional.profile.type.ts
export interface IProfessionalDocumentResponse {
  userId: string;
  email: string;
  name: string;
  role: string;
  bio: string;
  companyName: string;
  experience: number;
  profile_key: string;
  rating: number;
  sessionPrice: number;
  skills: string[];
  title: string;
  totalSessions: number;
  typesOfSessions: string[];
}